# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: آل ياسين لاكسسوار الموتال - نظام الكاشير
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/permissions:
    patch:
      operationId: updateUserPermissions
      tags: [users]
      summary: Update user permissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePermissionsBody"
      responses:
        "200":
          description: Updated permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, cashier]
        permissions:
          $ref: "#/components/schemas/Permissions"
        createdAt:
          type: string
      required:
        - id
        - username
        - name
        - role
        - permissions
        - createdAt
    Permissions:
      type: object
      properties:
        canViewDashboard:
          type: boolean
        canManageProducts:
          type: boolean
        canManageSuppliers:
          type: boolean
        canMakeSales:
          type: boolean
        canViewReports:
          type: boolean
        canManageUsers:
          type: boolean
        canManageDebts:
          type: boolean
        canViewPurchases:
          type: boolean
        canCreatePurchases:
          type: boolean
      required:
        - canViewDashboard
        - canManageProducts
        - canManageSuppliers
        - canMakeSales
        - canViewReports
        - canManageUsers
        - canManageDebts
        - canViewPurchases
        - canCreatePurchases
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, cashier]
      required:
        - username
        - password
        - name
        - role
    UpdateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, cashier]
    UpdatePermissionsBody:
      type: object
      properties:
        permissions:
          $ref: "#/components/schemas/Permissions"
      required:
        - permissions
...